# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for WartegKita — connecting customers with traditional Indonesian Warteg sellers.",
        "title": "WartegKita API",
        "contact": {
            "name": "WartegKita Team",
            "url": "https://github.com/Krisn4novianto/WartegKita_Project"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/address": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve saved delivery address for the authenticated user (requires Bearer JWT token)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User Address",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create or update delivery address for the authenticated user (requires Bearer JWT token)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Save User Address",
                "parameters": [
                    {
                        "description": "User address payload",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserAddress"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve user profile for the authenticated user (requires Bearer JWT token)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update user's name and email (requires Bearer JWT token)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update User Profile",
                "parameters": [
                    {
                        "description": "Profile update payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.UpdateProfileBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.UserAddress": {
            "type": "object",
            "properties": {
                "city_id": {
                    "type": "integer",
                    "example": 3201
                },
                "city_name": {
                    "type": "string",
                    "example": "Bogor"
                },
                "detail": {
                    "type": "string",
                    "example": "Jl. Merdeka No. 10"
                },
                "district_id": {
                    "type": "integer",
                    "example": 320101
                },
                "district_name": {
                    "type": "string",
                    "example": "Kecamatan Bogor Tengah"
                },
                "id": {
                    "type": "string",
                    "example": "018f4a12-89cd-7b1e-9a2c-3f4e56789abc"
                },
                "label": {
                    "type": "string",
                    "example": "Rumah"
                },
                "latitude": {
                    "type": "number",
                    "example": -6.597147
                },
                "longitude": {
                    "type": "number",
                    "example": 106.806038
                },
                "note": {
                    "type": "string",
                    "example": "Pagar besi"
                },
                "postal_code": {
                    "type": "string",
                    "example": "16111"
                },
                "province_id": {
                    "type": "integer",
                    "example": 32
                },
                "province_name": {
                    "type": "string",
                    "example": "Jawa Barat"
                },
                "user_id": {
                    "type": "string",
                    "example": "018f4a12-89cd-7b1e-9a2c-3f4e56789abc"
                }
            }
        },
...
        "routes.UpdateProfileBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "budi@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "Budi Santoso"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token as: Bearer \u003cyour_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}